debug

+++ packages/service/test/cli/compile-2.test.mjs
@@ -9,7 +9,7 @@ import { execa } from 'execa'
 import { fileURLToPath } from 'url'
 import { cliPath } from './helper.mjs'
 // import { cliPath, safeKill } from './helper.mjs'
-// import psList from 'ps-list'
+import psList from 'ps-list'
 
 process.setMaxListeners(100)
 
@@ -103,46 +103,45 @@ test('valid tsconfig file inside an inner folder', async (t) => {
   await cp(testDir, cwd, { recursive: true })
 
   try {
+    const start = new Date()
     const child = execa('node', [cliPath, 'compile'], {
       cwd,
       killSignal: 'SIGKILL'
     })
     console.log('5')
 
-    // const childPid = child.pid
+    const childPid = child.pid
     child.stdout.pipe(process.stdout)
     child.stderr.pipe(process.stderr)
 
+    const timeout = setInterval(async () => {
+      const processes = await psList()
+      const level1 = processes.filter((p) => p.pid === childPid)
+      const level2 = processes.filter((p) => level1.includes(p.ppid))
+      const level3 = processes.filter((p) => level2.includes(p.ppid))
+
+      console.log('level1', level1)
+      console.log('level2', level2)
+      console.log('level3', level3)
+
+      if (level1.length === 0) {
+        clearInterval(timeout)
+        return
+      }
+
+      if (new Date() - start > 60 * 1000) {
+        for (const p of level1) {
+          console.log('killing', p.pid)
+          child.kill('SIGKILL')
+          // await safeKill(p.pid)
+          // await execa('kill', ['-9', p.pid])
+          // await execa('taskkill', ['/pid', p.pid, '/f', '/t'])
+          // await execa('wmic', ['process', 'where', `ProcessId=${p.pid}`, 'terminate'])
+        }
+      }
+    }, 5000)
+
     await child
-
-    t.after(async () => {
-      child.kill('SIGKILL')
-    })
-
-    // const timeout = setInterval(async () => {
-    //   const processes = await psList()
-    //   const level1 = processes.filter((p) => p.pid === childPid)
-    //   const level2 = processes.filter((p) => level1.includes(p.ppid))
-    //   const level3 = processes.filter((p) => level2.includes(p.ppid))
-
-    //   console.log('level1', level1)
-    //   console.log('level2', level2)
-    //   console.log('level3', level3)
-
-    //   if (level1.length === 0) {
-    //     clearInterval(timeout)
-    //     return
-    //   }
-
-    //   for (const p of level1) {
-    //     console.log('killing', p.pid)
-    //     child.kill('SIGKILL')
-    //     // await safeKill(p.pid)
-    //     // await execa('kill', ['-9', p.pid])
-    //     // await execa('taskkill', ['/pid', p.pid, '/f', '/t'])
-    //     // await execa('wmic', ['process', 'where', `ProcessId=${p.pid}`, 'terminate'])
-    //   }
-    // }, 15000)
 
     // safeKill(parseInt(child.pid))
